+++ CHANGELOG.md
@@ -15,6 +15,8 @@ All notable changes to this project will be documented in this file.
   and using font embed dialog for fonts
 - Replacing morphshapes
 - SVG Export - stroke-bitmapId, fill-bitmapId attribute
+- Morphshape SVG Export - bitmap fill strokes
+- SVG Export/Import - retain bitmap fill smoothed attribute
 
 ### Fixed
 - [#1306], [#1768] Maximizing window on other than main monitor